Refactor images, add new summer party
