+++ app/routes/joblistings/components/JoblistingPage.tsx
@@ -5,7 +5,7 @@ import { useMemo } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { useLocation } from 'react-router-dom';
 import { fetchAll } from 'app/actions/JoblistingActions';
-import p6 from 'app/assets/sommerfest/6.png';
+import p6 from 'app/assets/sommerfest/9378543c80bd1b6a.png';
 import Sommerfest from 'app/components/Sommerfest/Sommerfest';
 import { selectJoblistings } from 'app/reducers/joblistings';
 import { useAppDispatch, useAppSelector } from 'app/store/hooks';
